Move tray handling from lib.rs to tray.rs
